Improved access control

+++ classes/Ijdb/Controllers/Joke.php
@@ -61,6 +61,12 @@ class Joke
     public function saveEdit()
     {
         $author = $this->authentication->getUser();
+        if (isset($_GET['id'])) {
+            $joke = $this->jokesTable->findById($_GET['id']);
+            if ($joke['authorid'] != $author['id']) {
+                return;
+            }
+        }
         $joke = $_POST['joke'];
         $joke['jokedate'] = new \DateTime();
         $joke['authorId'] = $author['id'];
@@ -70,6 +76,7 @@ class Joke
 
     public function edit()
     {
+        $author = $this->authentication->getUser();
         if (isset($_GET['id'])) {
             $joke = $this->jokesTable->findById($_GET['id']);
         }
@@ -78,7 +85,8 @@ class Joke
             'template' => 'editjoke.html.php',
             'title' => $title,
             'variables' => [
-                'joke' => $joke ?? null
+                'joke' => $joke ?? null,
+                'userId' => $author['id'] ?? null
             ]
         ];
     }
@@ -86,6 +94,11 @@ class Joke
 
     public function delete()
     {
+        $author = $this->authentication->getUser();
+        $joke = $this->jokesTable->findById($_POST['id']);
+        if ($joke['authorId'] != $author['id']) {
+            return;
+        }
         $this->jokesTable->delete($_POST['id']);
 
         header('location: /joke/list', 301);
